requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
attrs==25.1.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
ConfigArgParse==1.7
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
locust==2.33.0
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.2
playwright==1.50.0
pluggy==1.5.0
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
pytest==8.3.4
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-slugify==8.0.4
pyzmq==26.2.1
requests==2.32.3
setuptools==75.8.2
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
